Allow combination of MP and OutputBuffer plugins on Python 3.
You could not combine those two plugins in Python 3.4 (and presumably
any Python 3 version).
The error message when combining OutputBuffer and MP on Python 3.4
was internal error, 'state is not a dictionary'. It stems from
a python bug (https://bugs.python.org/issue26695) that prevents
unpickling objects which return a falsy value from __getstate__
when using _pickle.

The MP plugin used to send the Buffer object, and nested Streams, back
from the spawned processes, when a string would do just as well.

This patch moves the stream -> string conversion to setTestOutcome,
which still happens in the actual testing process.

It includes a test case exposing the issue.