Improve install docs for coverage-plugin
The coverage plugin docstrings should document the use of the
`nose2[coverage-plugin]` extra, not direct installation of `coverage`.
This way, the docs work for all versions.
Additionally, improve the ImportError warning message to properly
explain installation of `nose2[coverage-plugin]`.

Because this is a common point of confusion with extras, the example
in the module docstring notes how to specify the version of `nose2` at
the same time as the extra.

Resolves #355