Fix coverage reporting error: deferred too long
Waiting until startTestRun to start collecting coverage reporting
information is wrong! Upon closer inspection, it becomes obvious that if
coverage collection waits until then, it will miss any and all
import-time activity which happens while loading the testsuite. Since a
lot of tests are built off of real application code, that's a lot.

Adding this fix right away, and will try to write a testcase for it
later. Confirmed against another project's testsuite that this gets
*exactly* the same output as cov-core did.