Update setuptools invocation to handle coverage
Setuptools invocation runs differently from command-line usage. This is,
itself, a potential source of grief.
Try to make it behave more like `nose2` invocation by doing the
following:
- add ~/.{unittest,nose2}.cfg to loaded config
- trigger some events nose2 plugins want to use
- fix nose2.plugins.coverage to handle those events without handleArgs

Unfortunately, it looks like unifying these two modes of use more
would be a significant effort, and it's not obvious that the payoff is
worth the reward.

Resolves #317