Fix zipped egg discovery on newer setuptools
Avoid infinite recursion when loading tests from a zipped egg. Newer
setuptools can yield an empty-string when listing a dir in a zipped egg.
This looks like a setuptools bug, unless there's some subtlety to ""
being returned which makes it somehow valid. Even if it is a bug, it impacts
nose2 on newer setuptools versions.

Closes #383