Add "prettyassert" plugin, enabled by default
Make `assert` failures pretty, like `nosetests -d`.

If you call `nose2 --pretty-assert` or turn it on in config,
print the relevant exception statement, the exception message if given,
and attempt to resolve all names to their values by inspecting the
locals and globals available at the time of the exception to print
those as well.

The implementation of `nosetests -d` was the primary inspiration/guiding
force for this. The py.test approach is far too laborious to be worth
trying in nose2, as it involves rewriting bytecode on the fly, and its
original author noted that "Windows was a pain". Plus, there are all
sorts of caveats given with that implementation and scenarios where
it's known to blow up with other plugins, so we'd want this
implementation as a fallback for those cases anyway.

This is not an exact reimplementation of the `nosetests -d` approach,
however. I've tried to make some modest improvements to the behavior,
aided in part by the fact that we don't have to support ancient python
versions.