Resolve attributes in --pretty-assert
Given `assert x.y == 1`, `--pretty-assert` will now show values of `x`
and `x.y`.
This works correctly in most cases, including `(x).y`, `(x\n).y`, etc.

There are probably corner cases which it doesn't handle very well, and
it won't re-evaluate called functions.

It can be fooled in a number of scenarios like

>>> x = 1
>>> assert foo().x == 1

in which case it will wrongly think that `x` is a new name which needs
resolving, and grab `x` out of locals. The output is somewhat
surprising, perhaps, but it's unambiguous and not *wrong* in this case
("x" is 1, but that's not what we're interested in).