Fix --pretty-assert handling of multiline exprs
When an expression spans multiple lines, `--pretty-assert` could be
grabbing it in the middle of a line, thus giving it an unparseable
expression which fails tokenization.
To fix, keep adding lines until it tokenizes successfully. If it fails
at the end, mark it as such and change behavior appropriately. That
should never happen and I cannot find a case in which it does.

Also handle newlines correctly by skipping but doing nothing.

Add a test case which covers both attribute resolutions and multiline
asserts which require this re-parsing behavior to be handled correctly.