Make PrettyAssert ignore `raise AssertionError`
Ignore AssertionErrors which arise outside of the context of an `assert`
statement. This makes PrettyAssert *not* try to rewrite or alter the
errors raised by unittest assertion methods, which is a good thing.
It also cuts away some logic in the assertion statement search which was
meant to fall back to the failing line.
Add testing for this, and also for cases in which passing assertion
statements are not included in the output.