Allow plugins before ResultReporter to set status
If a plugin executes wasSuccessful() before the ResultReporter plugin,
ResultReporter would overwrite the result. To fix, tweak the
ResultSuccessEvent creation process to start at `None`, rather than
`False`, and treat `None` specially. This lets the ResultReporter
clearly see whether or not a result was explicitly set to `False` by a
prior plugin.

Also, fix the coverage plugin to be aware of this behavior when setting
a result status.

closes #395
closes #396