Remove default=None vals from sample plugin conf
If a plugin configuration parameter has a default of `None`, setting it
to `None` in the configuration sets it to the literal string, `'None'`.
To avoid this issue and the potential for confusion, omit anything with
`default=None` from the generated sample config.
This does not change the documentation of plugin parameters -- only the
config generated.

To make the phrasing more natural, slightly rearranges the checks done
in the sphinx extension.

closes #320