always-on plugins still allow their CLI switches
If you set `always-on` in a plugin config, we should still allow the
use of the relevant switch, but make it do nothing. However, we
should also use `argparse.SUPPRESS` to indicate that no help output
should be shown.

This way, config with many plugins loaded and always-on doesn't result
in `nose2 --help` being overly-long and confusing.

Add two tests for this in the pretty-assert tests.

fixes #432