Fix prettyassert truncating function defs
The prettyassert plugin was cutting off function definitions
incorrectly. They don't actually make things harder for us to tokenize
and process, so there's really no call for this.

Previously, the plugin was cutting off a line from the tops of frames.
This worked in testing because all the test functions had short
definitions with few parameters. But this is obviously not safe! Not
only will a multiline decorator blow things up, but so will long
parameter lists!

The solution is simple -- don't incorrectly truncate function
definitions. Cutting off the other end of the stack frame -- after the
current line -- is the part which is important and must remain.

fixes #433