mp: initiate new session based directly on old session
This takes advantage of the fact that a subprocess is a full copy of the
parent process, thus it has access to all the original data.

Rather than serializing the session information and passing it through
the pipe/connection, simply pass the original MultiProcess object to
procserver() as an argument.

Cons:
* Somewhat blurs the interface between a subprocess and its parent.

Pros:
* Less code.
* Eliminates the caveat that classes must be pickeable.

Continuing on this path should allow for parent-->child data passing
that is not currently available.