Update test config and switch to GHA
Drop Travis in favor of GitHub Actions.
Update the tox matrix to drop jython and py3.4, and to remove some
testing config for windows support (separate tox config, which is
never verified as working in Travis).

The unittest workaround file is converted into a platform agnostic
"self-test" wrapper script, and the comment is clarified to note that
the bug it addresses is in the stdlib multiprocessing package on
windows for py2.