Update version to 0.10.0 for release
Also update some more package metadata. Switch from author to
maintainer for info and point it at the dev@nose2.io listhost.

No offense is meant to JP, but he has not been actively involved in
the project for some time. For this reason, it no longer seems
reasonable for him to be listed in the package metadata.