Add dates to all releases in changelog
This provides a useful history of development and ensures that
everything is correctly annotated.
That includes untagged releases (0.1, 0.2) and removal of any releases
which were tagged but not published to pypi. Those are now listed
under the subsequent releases which include their changes.

The primary work of assembling dates was done via git commit dates
which may be misleading if a tag was added before or after the
release was actually published. Mostly, it was assembled with this
loop:

    git tag -l | while read t; do
      echo -e "$t ($(git log -n1 --format='%as' "$t"))\n------------------";
    done > dated-tags.txt

followed by manual curation and replacement.

For the untagged versions, the dates published were used.