Duplicate noqa comments for py38 AST fix
In py38, apparently a bug was fixed with the AST representation which
makes the logic for applying noqa comments to decorators variable by
python version. For detail, see:
  https://gitlab.com/pycqa/flake8/-/issues/583

Perhaps there's a more elegant fix, but a simple approach is to just
duplicate noqa comments needed in the tests.

With github actions enforcing that 3.9-flake8 is passing, it's
conceivable that python-version-specific flake8 behaviors like this
will lead to issues where commits pass CI and then fail on developers'
workstations. Those problems -- if they arise -- can be handled with
additional CI steps.