Handle datetime differences on old pythons
On older python versions, `datetime.datetime.fromisoformat` had not
been added yet. As a result, on these versions it cannot be used in
the testsuite to parse timestamp output from junitxml.

Only where it is not available, use `strptime` instead. If python2
support is ever dropped from nose2, this effectively documents how
the helper could be replaced with `fromisoformat`.