Update configs to drop py35, add 3.11-dev
tox, setuptools, and CI builds all updated.

py3.11-dev testing on linux is not skipped or marked
"continue-on-error". If it fails, we'll commit to fixing it, so that
`nose2` can be used to test against the latest. If 3.11-dev is ever
really broken for a long time, this can be removed.