Switch to vendored copy of six
`six` is written as a single module in order to enable very easy
(manual) vendoring. Switching to a vendored copy has a couple of
benefits:
- installation becomes faster and simpler with one fewer dependency
  (this gain is probably marginal)
- the application under test may install `six`, and may use any
  version of it, including ones which are not compatible with `nose2`

The second point is valuable in some cases. It means that it will no
longer be possible to install `nose2` and then break it accidentally by
installing an older version of `six`.