Remove install_requires dependency on 'coverage'
This can be traced back to an added dependency on code-cov (prior to
coverage) which doesn't appear to have been correct. It was added to
the installation requirements rather than the dev requirements.
It should be possible to use ``nose2`` in an environment without
``coverage`` installed (but not to use the coverage plugin).

As a result, `install_requires` can be removed entirely.