Comment out flake8 max-complexity setting
This value appears to have been reached experimentally, so we should
preserve the learned value.

But it hasn't been discussed/evaluated against the nose2 codebase, and
needs some thought before it can be included. In order to unblock
other config changes, comment it out and move it to the bottom of the
flake8 config.