Make --junit-xml-path implicitly register plugin
Rather than requiring `--junit-xml --junit-xml-path=a.xml` in order to
write to a file, make `--junit-xml-path=a.xml` sufficient.

Even when the plugin is not registered, it can still have its argument
handler trigger, giving us an opportunity to trivially call
`self.register()` (which is what happens if the activation switch
flag is passed).

Add a test case + changelog entry for this.

closes #521