Cleanup attribute inspection
1. Replace `util.isgenerator` with `inspect.isgeneratorfunction`

The `util.isgenerator` check allowed for a `testGenerator` attribute
to indicate that a method is a generator. However, this doesn't appear
to be used ever and there are no test cases in the testsuite which
exercise this functionality. Remove the unnecessary check.

2. Retrofit most `dir()` usage with `util.iter_attrs`

Several code locations iterate over object, module, and class
attributes using `dir()` + `getattr()`. `dir()` is necessary since
`__slots__` can make `vars()` unreliable for attribute iteration, and
this subtlety could easily be lost in a refactor or change in the
future. Consolidate the logic into a single helper which is used in
all locations (except those which only want names, not attribute
values).