Move to pyproject.toml for setuptools data (#572)
This requires more modern versions of setuptools to build, but results
in declarative metadata with fewer pieces of disparate config.

The setup.py stub is kept for now to reduce the number of steps
redistributors of nose2 will have to take to adapt. It can be removed
in a subsequent release.

Also, remove a defunct `graft` rule from `MANIFEST.in`.
Apply a fix from pre-commit hooks.