Fix verbose reporting of skipped tests
On 3.12.1+, unittest doesn't call `startTest` for skipped tests. This
is treated as a fix to how tests are counted, which is why it appeared
in a point release.

Because the nose2 path for test reporting uses the test result methods
as the point of connection between `unittest` and nose2 plugins,
losing the `startTest` call in this case means that the reporter
plugin doesn't emit proper output.

The test result object now tracks whether or not a test has been
started, and the reporter will check this attribute to ensure that the
skipped test output is correct.