Move back to flake8, remove ruff
Choice of linter toolchain is a matter for project developers. `ruff`
was pushed by an external contributor, and was accepted to hopefully
improve interest/engagement in maintenance.

However, there was no particular reason to use `ruff` and it puts us
in the position of opting out of a lot of configs rather than opting
in to what we want.

flake8 also offers a plugin framework which allows for checks like
`flake8-typing-as-t` to be easily added post-hoc.

In addition to flake8, reapply isort and pyupgrade, and add slyp.