Add runInProc to testsuite, fixing coverage tests
nose2's testsuite does not really create subprocesses for most
executions. This keeps things fast by doing all of the work in the
current process, but is self-defeating when we really *do* want to
isolate the work into a proper sub-execution and check the results, as
is the case with `coverage` invocations (which involve coverage's own
state on a per-interpreter basis).

To resolve, add a helper for doing true subprocess invocations where
appropriate, and use it to fix the broken coverage tests.

A few minor edits were also needed to get this all working:
- fix some subproc invocations using `target=...`
- remove strange code from the "coverage of imports" scenario [^1]

[^1]: I was the author of said scenario, many years ago. The commit
history doesn't show it, but my suspicion is that with the test
skipped, it was just a save of whatever my work in progress was, as I
tried to get the test working at the time. It's hard to be certain,
but the code in that file doesn't have a clearly credible reason to
exist.