Convert 'toml' usage to tomllib/tomli
tomli was ported to stdlib tomllib, so it's preferable to use these
implementations for clearer (and more standard/common style)
compatibility across versions.

In order to have the failthrough behavior be clear, with dispatch
onto TOML support controlled via a bool, this is wrapped in a module
which provides failover on import errors per the normal path to
prefer tomllib where available.

'tomli' is not required by nose2, not even on lower python versions,
to ensure that application testing which is sensitive to and
dispatched over the presence/absence of a toml loader is supported.

tox matrix config supports toml-enabled test invocation for the lower
python versions.