Fix a bug with ini configs lacking '.cfg' suffix
The check for which config loading path to use incorrectly treated
'.cfg' as the only valid suffix for ini configs.
This is not the case for a config loaded with `-c`.

To fix, introduce a regression test, then make some strategic use of
pathlib to simplify path/file inspection and fix the issue.

fixes #614